requirements.txt
Flask==0.9
Flask-Bcrypt==0.5.2
Flask-Login==0.1.3
Jinja2==2.6
PennSDK==0.1.3
PyAudio==0.2.7
PyRSS2Gen==1.0.0
Pygments==1.6
Sphinx==1.2b3
Twisted==12.0.0
Werkzeug==0.8.3
altgraph==0.9
bdist-mpkg==0.4.4
beautifulsoup4==4.1.3
bonjour-py==0.3
docutils==0.11
flake8==1.7.0
geopy==0.95.1
googlemaps==1.0.2
macholib==1.4.2
modulegraph==0.9.1
node==0.9.11
nose==1.2.1
numpy==1.7.0
odict==1.5.1
plumber==1.2
py-bcrypt==0.4
py2app==0.6.3
pyOpenSSL==0.13
pyobjc-core==2.3.2a0
pyobjc-framework-AddressBook==2.3.2a0
pyobjc-framework-AppleScriptKit==2.3.2a0
pyobjc-framework-AppleScriptObjC==2.3.2a0
pyobjc-framework-Automator==2.3.2a0
pyobjc-framework-CFNetwork==2.3.2a0
pyobjc-framework-CalendarStore==2.3.2a0
pyobjc-framework-Cocoa==2.3.2a0
pyobjc-framework-Collaboration==2.3.2a0
pyobjc-framework-CoreData==2.3.2a0
pyobjc-framework-CoreLocation==2.3.2a0
pyobjc-framework-CoreText==2.3.2a0
pyobjc-framework-DictionaryServices==2.3.2a0
pyobjc-framework-ExceptionHandling==2.3.2a0
pyobjc-framework-FSEvents==2.3.2a0
pyobjc-framework-InputMethodKit==2.3.2a0
pyobjc-framework-InstallerPlugins==2.3.2a0
pyobjc-framework-InstantMessage==2.3.2a0
pyobjc-framework-InterfaceBuilderKit==2.3.2a0
pyobjc-framework-LatentSemanticMapping==2.3.2a0
pyobjc-framework-LaunchServices==2.3.2a0
pyobjc-framework-Message==2.3.2a0
pyobjc-framework-OpenDirectory==2.3.2a0
pyobjc-framework-PreferencePanes==2.3.2a0
pyobjc-framework-PubSub==2.3.2a0
pyobjc-framework-QTKit==2.3.2a0
pyobjc-framework-Quartz==2.3.2a0
pyobjc-framework-ScreenSaver==2.3.2a0
pyobjc-framework-ScriptingBridge==2.3.2a0
pyobjc-framework-SearchKit==2.3.2a0
pyobjc-framework-ServerNotification==2.3.2a0
pyobjc-framework-ServiceManagement==2.3.2a0
pyobjc-framework-SyncServices==2.3.2a0
pyobjc-framework-SystemConfiguration==2.3.2a0
pyobjc-framework-WebKit==2.3.2a0
pyobjc-framework-XgridFoundation==2.3.2a0
python-dateutil==1.5
redis==2.8.0
requests==1.2.3
simplejson==3.1.3
vboxapi==1.0
virtualenv==1.9.1
wsgiref==0.1.2
xattr==0.6.2
zope.deprecation==4.0.2
zope.event==4.0.2
zope.interface==3.5.1
zope.lifecycleevent==4.0.3
